Moved func_api.py into Auxiliary module
